+++ component/http/handler_test.go
@@ -8,7 +8,6 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/beatlabs/patron/correlation"
 	"github.com/beatlabs/patron/encoding"
 	"github.com/beatlabs/patron/encoding/json"
 	"github.com/beatlabs/patron/encoding/protobuf"
@@ -120,33 +119,6 @@ func request(t *testing.T, contentType, accept string) *http.Request {
 		req.Header.Set(encoding.AcceptHeader, accept)
 	}
 	return req
-}
-
-func Test_getOrSetCorrelationID(t *testing.T) {
-	t.Parallel()
-	withID := http.Header{correlation.HeaderID: []string{"123"}}
-	withoutID := http.Header{correlation.HeaderID: []string{}}
-	withEmptyID := http.Header{correlation.HeaderID: []string{""}}
-	missingHeader := http.Header{}
-	type args struct {
-		hdr http.Header
-	}
-	tests := map[string]struct {
-		args args
-	}{
-		"with id":        {args: args{hdr: withID}},
-		"without id":     {args: args{hdr: withoutID}},
-		"with empty id":  {args: args{hdr: withEmptyID}},
-		"missing Header": {args: args{hdr: missingHeader}},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			assert.NotEmpty(t, correlation.GetOrSetHeaderID(tt.args.hdr))
-			assert.NotEmpty(t, tt.args.hdr[correlation.HeaderID][0])
-		})
-	}
 }
 
 func Test_handleSuccess(t *testing.T) {
