+++ correlation/correlation_test.go
@@ -2,6 +2,7 @@ package correlation
 
 import (
 	"context"
+	"net/http"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -34,4 +35,31 @@ func TestContextWithID(t *testing.T) {
 	val, ok := ctx.Value(idKey).(string)
 	assert.True(t, ok)
 	assert.Equal(t, "123", val)
+}
+
+func TestGetOrSetHeaderID(t *testing.T) {
+	t.Parallel()
+	withID := http.Header{HeaderID: []string{"123"}}
+	withoutID := http.Header{HeaderID: []string{}}
+	withEmptyID := http.Header{HeaderID: []string{""}}
+	missingHeader := http.Header{}
+	type args struct {
+		hdr http.Header
+	}
+	tests := map[string]struct {
+		args args
+	}{
+		"with id":        {args: args{hdr: withID}},
+		"without id":     {args: args{hdr: withoutID}},
+		"with empty id":  {args: args{hdr: withEmptyID}},
+		"missing Header": {args: args{hdr: missingHeader}},
+	}
+	for name, tt := range tests {
+		tt := tt
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			assert.NotEmpty(t, GetOrSetHeaderID(tt.args.hdr))
+			assert.NotEmpty(t, tt.args.hdr[HeaderID][0])
+		})
+	}
 }
